refactor: RoundProgressActivity toolbar 대체
뒤로 버튼 눌렀을 때 로직도 추가
